CuBoulder/tiamat-theme#266 adds search placeholder setting; splits off "pages & search" into a separate tab
