+++ ucb_site_configuration.links.task.yml
@@ -8,18 +8,23 @@ ucb_site_configuration.appearance_form:
   title: 'Appearance'
   base_route: ucb_site_configuration
   weight: 1
+ucb_site_configuration.pages_form:
+  route_name: ucb_site_configuration.pages_form
+  title: 'Pages & search'
+  base_route: ucb_site_configuration
+  weight: 2
 ucb_site_configuration.contact_info_form:
   route_name: ucb_site_configuration.contact_info_form
   title: 'Contact info'
   base_route: ucb_site_configuration
-  weight: 2
+  weight: 3
 ucb_site_configuration.related_articles_form:
   route_name: ucb_site_configuration.related_articles_form
   title: 'Related articles'
   base_route: ucb_site_configuration
-  weight: 3
+  weight: 4
 ucb_site_configuration.external_services:
   route_name: entity.ucb_external_service_include.collection
   title: 'Third-party services'
   base_route: ucb_site_configuration
-  weight: 4
+  weight: 5
